requirements.txt
requests
BeautifulSoup4
pycrypto
future
